feat(cart): seat picker dialog feeds the cart (#853)
Extracts the user_choice seat-picking UI out of SeatAssignmentSection into
SeatPickerPanel (map/list views, floor/scope/view toggles, price legend,
409-conflict live region, cross-sector switch) so there is one implementation
consumed by both the legacy dialog and the new cart-flow SeatPickerDialog.

SeatPickerDialog hosts the panel over a transient SeatHoldController
constructed fresh per picking session, resolving the chicken-and-egg problem
where a cart group's own controller doesn't exist until seats have been
picked. Seeds/adopts from the shared query cache (warm from any other
controller on the event), hands picked seats to the cart on Done
(cart.setSeatIds), and releases un-handed-off holds on close/unmount.

TierCard gains a "Pick seats…" entry point plus a held-seats summary chip
for user_choice tiers when a cart is available, threaded through
TicketTierList's new onPickSeats callback.

Adds cart.pickSeatsTitle, cart.pickSeats, cart.seatsPicked (plural), and
cart.seatPickerDone across all six locales.
