+++ messages/de.json
@@ -9502,7 +9502,20 @@
 		"summaryRegion": "Warenkorb-Übersicht",
 		"discountLink": "Rabattcode?",
 		"seatsHeld": "Plätze reserviert",
-		"holdsExpired": "Deine reservierten Plätze sind abgelaufen — bitte wähle sie erneut aus."
+		"holdsExpired": "Deine reservierten Plätze sind abgelaufen — bitte wähle sie erneut aus.",
+		"pickSeatsTitle": "Plätze wählen — {tierName}",
+		"pickSeats": "Plätze wählen…",
+		"seatsPicked": [
+			{
+				"declarations": ["input count", "local countPlural = count: plural"],
+				"selectors": ["countPlural"],
+				"match": {
+					"countPlural=one": "{count} Platz · bearbeiten",
+					"countPlural=*": "{count} Plätze · bearbeiten"
+				}
+			}
+		],
+		"seatPickerDone": "Fertig"
 	},
 	"cartSheet": {
 		"title": "Kasse",
