+++ messages/en.json
@@ -9502,7 +9502,20 @@
 		"summaryRegion": "Cart summary",
 		"discountLink": "Discount code?",
 		"seatsHeld": "Seats held",
-		"holdsExpired": "Your held seats expired — please pick them again"
+		"holdsExpired": "Your held seats expired — please pick them again",
+		"pickSeatsTitle": "Pick seats — {tierName}",
+		"pickSeats": "Pick seats…",
+		"seatsPicked": [
+			{
+				"declarations": ["input count", "local countPlural = count: plural"],
+				"selectors": ["countPlural"],
+				"match": {
+					"countPlural=one": "{count} seat · edit",
+					"countPlural=*": "{count} seats · edit"
+				}
+			}
+		],
+		"seatPickerDone": "Done"
 	},
 	"cartSheet": {
 		"title": "Checkout",
