+++ messages/pt.json
@@ -9502,7 +9502,20 @@
 		"summaryRegion": "Resumo do carrinho",
 		"discountLink": "Código de desconto?",
 		"seatsHeld": "Lugares reservados",
-		"holdsExpired": "Os teus lugares reservados expiraram — volta a selecioná-los."
+		"holdsExpired": "Os teus lugares reservados expiraram — volta a selecioná-los.",
+		"pickSeatsTitle": "Escolher lugares — {tierName}",
+		"pickSeats": "Escolher lugares…",
+		"seatsPicked": [
+			{
+				"declarations": ["input count", "local countPlural = count: plural"],
+				"selectors": ["countPlural"],
+				"match": {
+					"countPlural=one": "{count} lugar · editar",
+					"countPlural=*": "{count} lugares · editar"
+				}
+			}
+		],
+		"seatPickerDone": "Concluído"
 	},
 	"cartSheet": {
 		"title": "Pagamento",
